reduce number of events on code refresh
We were calling refresh_timer from every single code in the list. We
only need to call it once, and then only if the creds list contents have
changed.

Logically, the list only needs to be rebuilt if the contents have
changed. But we want a separate event to indicate when the codes have
finished being generated.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>